refactor: ui와 데이터 관리 영역 분리
